Re-arrange sigproc code to make it compatible with CCS C89 restrictions
C99 standardised in May 2000, it's 11.5 years old and still TI haven't adopted
it...

git-svn-id: svn+ssh://svn.forge.ecs.soton.ac.uk/projects/sanchp3yp@100 bc9da1e1-76ae-0410-a611-e94dc21032b9

+++ work/code/sigproc.h
@@ -6,8 +6,8 @@
 
 #define CROSS_CORR_MAX_DELAY 200
 
-void cancel(void * ring1, int16_t * ptr1in, void * ring2, int16_t *ptr2in, void * res, int16_t * ptrres);
-void crosscorr(void * ring1, int16_t * ptr1in, void * ring2, int16_t * ptr2in, double * res, int length, int maxdel);
+void cancel(void * ring1, int32_t * ptr1in, void * ring2, int32_t *ptr2in, void * res, int32_t * ptrres);
+void crosscorr(void * ring1, int32_t * ptr1in, void * ring2, int32_t * ptr2in, double * res, int length, int maxdel);
 int dpeak(double * arr, int length);
 double mean(void * ring, int32_t * ptrin);
